Move additional data type to properties types file
