+++ src/components/Property/types.ts
@@ -6,25 +6,6 @@ export interface PropertyProps {
   readOnly: boolean;
 }
 
-export interface AdditionalDataVectorMatrix {
-  Exponent: number; // TODO: handle the exponent
-  MaximumValue: number[];
-  MinimumValue: number[];
-  SteppingValue: number[];
-}
-
-export interface AdditionalDataNumber {
-  Exponent: number;
-  MaximumValue: number;
-  MinimumValue: number;
-  SteppingValue: number;
-}
-
-export type ViewOptionsVector = {
-  Color?: boolean;
-  MinMaxRange?: boolean;
-};
-
 // In OpenSpace the options are represented like so:
 // { 0: "Option 1"}, { 1: "Option 2"}
 // The key is a number but in a string format.
@@ -39,3 +20,28 @@ export interface AdditionalDataOptions {
 export interface AdditionalDataSelection {
   Options: string[];
 }
+
+export interface AdditionalDataVectorMatrix {
+  Exponent: number; // TODO: handle the exponent
+  MaximumValue: number[];
+  MinimumValue: number[];
+  SteppingValue: number[];
+}
+
+export interface AdditionalDataNumber {
+  Exponent: number;
+  MaximumValue: number;
+  MinimumValue: number;
+  SteppingValue: number;
+}
+
+export type AdditionalData =
+  | AdditionalDataNumber
+  | AdditionalDataVectorMatrix
+  | AdditionalDataOptions
+  | AdditionalDataSelection;
+
+export type ViewOptionsVector = {
+  Color?: boolean;
+  MinMaxRange?: boolean;
+};
