+++ src/components/PropertyOwner/PropertyOwnerContent.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@mantine/core';
 
 import { Property } from '@/components/Property/Property';
-import { usePropertyOwner, useGetVisibleProperties } from '@/hooks/propertyOwner';
+import { useGetVisibleProperties, usePropertyOwner } from '@/hooks/propertyOwner';
 import { GlobeLayersPropertyOwner } from '@/panels/Scene/GlobeLayers/GlobeLayersPropertyOwner';
 import { useAppSelector } from '@/redux/hooks';
 import { Uri } from '@/types/types';
