+++ src/panels/Scene/SceneGraphNode/SceneGraphNodeView.tsx
@@ -2,7 +2,7 @@ import { Box, Tabs, Text, Tooltip } from '@mantine/core';
 
 import { PropertyOwner } from '@/components/PropertyOwner/PropertyOwner';
 import { PropertyOwnerContent } from '@/components/PropertyOwner/PropertyOwnerContent';
-import { usePropertyOwner, useGetVisibleProperties } from '@/hooks/propertyOwner';
+import { useGetVisibleProperties, usePropertyOwner } from '@/hooks/propertyOwner';
 import { useAppSelector } from '@/redux/hooks';
 import { TransformType } from '@/types/enums';
 import { Uri } from '@/types/types';
