+++ src/redux/propertytree/propertyTreeMiddleware.ts
@@ -1,11 +1,11 @@
 import { createAction, createAsyncThunk } from '@reduxjs/toolkit';
 
 import { api } from '@/api/api';
+import { AdditionalData } from '@/components/Property/types';
 import { onOpenConnection } from '@/redux/connection/connectionSlice';
 import { refreshGroups } from '@/redux/groups/groupsSliceMiddleware';
 import type { AppStartListening } from '@/redux/listenerMiddleware';
 import {
-  AdditionalData,
   Identifier,
   Properties,
   Property,
